Har skilt kode for web og app i forskjellige servicer
